pyproject.toml
[tool.poetry]
name = "coachdiary"
version = "0.1.0"
description = ""
authors = ["Ivan Studinsky <ivan.studinsky@saritasa.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
invoke = "^2.2.0"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-soft-delete = "^1.0.12"
drf-writable-nested = "^0.7.0"
django-dirtyfields = "^1.9.2"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
